requirements.txt
alabaster==0.7.12
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.4
async-timeout==3.0.1
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
clarifai==2.6.2
colorama==0.4.3
configparser==3.8.1
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.3
django-advanced-filters==1.3.0
django-allauth==0.44.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==1.1.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.7.1
docutils==0.17.1
EasyProcess==0.3
emoji==0.6.0
enum-compat==0.0.3
et-xmlfile==1.0.1
filelock==3.0.12
fsspec==0.8.5
future==0.18.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
imagesize==1.2.0
instapy==0.6.13
jdcal==1.4.1
Jinja2==3.0.1
jsonschema==2.6.0
MarkupSafe==2.0.1
MeaningCloud-python==2.0.0
multidict==5.1.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.1
numpy==1.19.1
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.9
panda==0.3.1
pandas==1.2.3
paramiko==2.7.2
plyer==2.0.0
protobuf==3.14.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pysftp==0.2.9
python-dateutil==2.8.1
python-telegram-bot==13.1
python3-openid==3.2.0
pytz==2020.1
PyVirtualDisplay==2.0
PyYAML==5.4.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
serialization==1.2.3
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.1
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
tornado==6.1
tqdm==4.56.0
twilio==6.45.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.4.6
webdriverdownloader==1.1.0.3
wget==3.2
xlrd==2.0.1
zope.interface==5.4.0
